backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Commented out for Windows compatibility
alembic==1.13.1
openai==1.3.7
# chromadb==0.4.18  # Commented out for Windows compatibility
# pymupdf==1.23.8  # Commented out for Windows compatibility
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
